requirement.txt
certifi==2023.7.22
charset-normalizer==3.2.0
docopt==0.6.2
et-xmlfile==1.1.0
freetype-py==2.4.0
idna==3.4
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
yarg==0.1.9
